fix: revert servings data type to Double in MealFoodRequest
Changes:
- Changed servings field type back from Integer to Double in MealFoodRequest DTO:
  - @Positive validation constraint remains appropriate for Double values
  - Double allows fractional servings (e.g., 0.5, 1.5 servings)
  - More realistic for food serving measurements which can be fractional
- This aligns with the MealFoodMapper implementation that expects Double servings

This change supports more precise serving measurements for nutritional calculations.

+++ src/main/java/com/dev/quikkkk/dto/request/MealFoodRequest.java
@@ -18,5 +18,5 @@ public class MealFoodRequest {
     private String foodId;
 
     @Positive(message = "VALIDATION.MEAL_FOOD.SERVINGS.POSITIVE")
-    private Integer servings;
+    private Double servings;
 }
